WIP: create a separate logout page, clearing all logout made inline in js navigation
TODO:
- ensure that the navbar is updated to the latest markup, so that the logout link query is accurate
- remove previous logout functions
- implement, logout confirm action
- etc.

+++ modules/core/handler_modules.php
@@ -774,6 +774,14 @@ class Hm_Handler_logout extends Hm_Handler_Module {
      * Clean up everything on logout
      */
     public function process() {
+        if ($this->request->get['prompt'] ?? false) {
+            $backQuery = isset($this->request->get['back_query']) ? unserialize(base64_decode($this->request->get['back_query'])): [];
+
+            $this->out('cancel_logout_url', '?' . http_build_query($backQuery));
+            
+            return;
+        }
+
         if (array_key_exists('logout', $this->request->post) && !$this->session->loaded) {
             $this->session->destroy($this->request);
             Hm_Msgs::add('Session destroyed on logout', 'info');
